pyproject.toml
[project]
name = "openspp-deployment-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docker>=7.1.0",
    "gitpython>=3.1.45",
    "invoke>=2.2.0",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.47.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
]
